// sourcecode/BTrak.Packages/snova-payroll-widgets/module/module-components/src/lib/payrollwidgetsandreports/models/financialyearconfigurationsmodel.ts
export class FinancialYearConfigurationsModel{
    financialYearConfigurationsId: string;
    fromMonth: number;
    toMonth: number;
    isArchived: boolean;
    timeStamp: any;
    countryId: string;
    countryName: string;
    financialYearTypeId: string;
    financialYearTypeName: string;
    activeFrom: Date;
    activeTo: Date;
}